fix(pegboard): cast grade::int in ListByClubAsync for PlayerClubView mapping
Positional-record Dapper mapping is strict about smallint(Int16) vs int?(Int32);
the new grade column broke club roster reads (500). Matches the documented cast seam.

+++ Repositories/PlayerRepository.cs
@@ -75,7 +75,7 @@ public sealed class PlayerRepository : IPlayerRepository
         using var conn = _factory.Create();
         var rows = await conn.QueryAsync<PlayerClubView>(
             new CommandDefinition(
-                @"SELECT p.id AS player_id, p.full_name, p.gender, pc.type, p.grade
+                @"SELECT p.id AS player_id, p.full_name, p.gender, pc.type, p.grade::int
                   FROM player_clubs pc
                   JOIN players p ON p.id = pc.player_id
                   WHERE pc.club_id = @clubId
